% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bushfire.weights}
\alias{bushfire.weights}
\title{Weights for Bushfire scars.}
\format{A vector of length 38.}
\usage{
bushfire.weights
}
\description{
The bushfire data set was used by Campbell (1984, 1989) to locate bushfire scars.
The dataset contains satelite measurements on five frequency bands, corresponding
to each of 38 pixels.
}
\details{
For testing purposes, \code{bushfire.weights} provides artificial weights created
according to: \code{bushfire.weights <- rep(c(1,2,5), length = nrow(bushfire))}
}
\examples{
data(bushfire.weights)
}
\references{
Campbell, N. (1989) Bushfire Mapping using NOAA AVHRR Data. Technical Report.
Commonwealth Scientific and Industrial Research Organisation, North Ryde.
}
\keyword{datasets}
