% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modi-internal.R
\name{EM.normal}
\alias{EM.normal}
\title{EM for multivariate normal data}
\usage{
EM.normal(
  data,
  weights = rep(1, nrow(data)),
  n = sum(weights),
  p = ncol(data),
  s.counts,
  s.id,
  S,
  T.obs,
  start.mean = rep(0, p),
  start.var = diag(1, p),
  numb.it = 10,
  Estep.output = FALSE
)
}
\arguments{
\item{data}{matrix or dataframe with data.}

\item{weights}{vector of weights.}

\item{n}{number of rows.}

\item{p}{number of columns.}

\item{s.counts}{s.counts.}

\item{s.id}{s.id.}

\item{S}{S.}

\item{T.obs}{T.obs.}

\item{start.mean}{initial center.}

\item{start.var}{initial variance.}

\item{numb.it}{numb.it.}

\item{Estep.output}{Estep.output.}
}
\description{
This version of EM does not contain the computation of the
observed sufficient statistics, they will be computed in the
main program of BEM and passed as parameters as well as the
statistics on the missingness patterns.
}
\author{
Beat Hulliger
}
\keyword{internal}
