% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modi-internal.R
\name{ind.dij}
\alias{ind.dij}
\title{Addressing function for Epidemic Algorithm}
\usage{
ind.dij(i, j, n)
}
\arguments{
\item{i}{index i.}

\item{j}{index j.}

\item{n}{number of rows.}
}
\description{
Utility function for Epidemic Algorithm.
}
\author{
Cédric Béguin
}
\keyword{internal}
