% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modi-internal.R
\name{ind.dijs}
\alias{ind.dijs}
\title{Addressing function for Epidemic Algorithm}
\usage{
ind.dijs(i, js, n)
}
\arguments{
\item{i}{index i.}

\item{js}{indexes js.}

\item{n}{number of rows.}
}
\description{
Utility function for Epidemic Algorithm.
}
\author{
Cédric Béguin
}
\keyword{internal}
