% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modi-package.R
\docType{package}
\name{modi-package}
\alias{modi-package}
\title{modi: Multivariate Outlier Detection and Imputation for Incomplete Survey Data}
\description{
Algorithms for multivariate outlier detection when missing values occur. Algorithms are based on Mahalanobis distance or data depth. Imputation is based on the multivariate normal model or uses nearest neighbour donors. The algorithms take sample designs, in particular weighting, into account. The methods are described in Bill and Hulliger (2016) \doi{10.17713/ajs.v45i1.86}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/martinSter/modi}
  \item Report bugs at \url{https://github.com/martinSter/modi/issues}
}

}
\author{
\strong{Maintainer}: Beat Hulliger \email{beat.hulliger@fhnw.ch}

Other contributors:
\itemize{
  \item Martin Sterchi \email{martin.sterchi@fhnw.ch} [contributor]
  \item Tobias Schoch \email{tobias.schoch@fhnw.ch} [contributor]
}

}
\keyword{internal}
