% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modi-internal.R
\name{nz.min}
\alias{nz.min}
\title{Non-zero non-missing minimum function}
\usage{
nz.min(x)
}
\arguments{
\item{x}{vector of data.}
}
\description{
Returns the non-zero non-missing minimum.
}
\author{
Beat Hulliger
}
\keyword{internal}
