% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modi-internal.R
\name{psi.lismi}
\alias{psi.lismi}
\title{psi-function}
\usage{
psi.lismi(d, present, psi.par = c(2, 1.25))
}
\arguments{
\item{d}{vector of distances.}

\item{present}{present.}

\item{psi.par}{parameters.}
}
\description{
Defined in Little and Smith for ER algorithm
}
\author{
Beat Hulliger
}
\keyword{internal}
