% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modi-internal.R
\name{sweep.operator}
\alias{sweep.operator}
\title{Sweep operator}
\usage{
sweep.operator(M, k, reverse = FALSE)
}
\arguments{
\item{M}{a matrix.}

\item{k}{column.}

\item{reverse}{either \code{TRUE} or \code{FALSE}.}
}
\description{
Definition of the sweep and reverse-sweep operator (Schafer pp 159-160)
}
\author{
Beat Hulliger
}
\keyword{internal}
