% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modi-internal.R
\name{weightsum}
\alias{weightsum}
\title{Utility function for TRC.R among others}
\usage{
weightsum(observations, weights, value, lt = TRUE)
}
\arguments{
\item{observations}{vector of observations.}

\item{weights}{vector of weights.}

\item{value}{value.}

\item{lt}{either \code{TRUE} or \code{FALSE}.}
}
\description{
Sum of weights for observations < value (if lt = TRUE)
or observations=value (if lt = FALSE).
}
\author{
Beat Hulliger
}
\keyword{internal}
